api-gateway/requirements.txt
fastapi
uvicorn
aio_pika
redis
python-dotenv
